tor-netdoc: Remove unnecessary casts

+++ crates/tor-netdoc/src/doc/netstatus.rs
@@ -1274,7 +1274,7 @@ impl SignatureGroup {
             }
         }
 
-        ok.len() > (n_authorities / 2) as usize
+        ok.len() > (n_authorities / 2)
     }
 }
